+++ packages/core/src/m3/builtins.ts
@@ -4,6 +4,7 @@ import { PropertyValueSerializer } from "../serializer.js"
 import { currentReleaseVersion } from "../version.js"
 import { LanguageFactory } from "./factory.js"
 import { Classifier, Concept, DataType, lioncoreBuiltinsKey, Property } from "./types.js"
+import { unresolved } from "../references.js"
 
 const lioncoreBuiltinsIdAndKeyGenerator: StringsMapper = (...names) => [lioncoreBuiltinsKey, ...names.slice(1)].join("-")
 
@@ -128,7 +129,7 @@ export class BuiltinPropertyValueDeserializer
             throw new Error(`can't deserialize undefined as the value of required property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}")`)
         }
         const { type } = property
-        if (type == null) {
+        if (type === unresolved) {
             throw new Error(`can't deserialize property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}") with unspecified type`)
         }
         const specificDeserializer = this.byType(type)
@@ -171,7 +172,7 @@ export class BuiltinPropertyValueSerializer extends DataTypeRegister<(value: unk
             throw new Error(`can't serialize undefined as the value of required property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}")`)
         }
         const { type } = property
-        if (type == null) {
+        if (type === unresolved) {
             throw new Error(`can't serialize property "${property.name}" (on classifier "${property.classifier.name}" in language "${property.classifier.language.name}") with unspecified type`)
         }
         const specificSerializer = this.byType(type)
@@ -194,3 +195,4 @@ export const builtinPropertyValueSerializer = new BuiltinPropertyValueSerializer
 export class DefaultPrimitiveTypeSerializer extends BuiltinPropertyValueSerializer {}
 
 export { builtinPrimitives, builtinClassifiers, builtinFeatures, isBuiltinNodeConcept, lioncoreBuiltins, shouldBeIdentical }
+