+++ packages/core/src/references.ts
@@ -1,5 +1,3 @@
-import { LionWebId } from "@lionweb/json"
-
 import { Node } from "./types.js"
 
 
@@ -14,12 +12,12 @@ export const unresolved = null
  * Note: this type is primarily meant to be used to type nodes’ properties,
  * but should be avoided as a return type for “auxiliary” functions.
  */
-export type SingleRef<T> = typeof unresolved | T
+export type SingleRef<NT extends Node> = typeof unresolved | NT
 
 /**
  * @return whether a given (at most) single-valued reference actually refers to something.
  */
-export const isRef = <T>(ref?: SingleRef<T>): ref is T =>
+export const isRef = <NT extends Node>(ref?: SingleRef<NT>): ref is NT =>
     ref !== undefined && ref !== unresolved
 
 /**
@@ -27,18 +25,5 @@ export const isRef = <T>(ref?: SingleRef<T>): ref is T =>
  * Note: this type is primarily meant to be used to type nodes’ properties,
  * but should be avoided as a return type for “auxiliary” functions.
  */
-export type MultiRef<T> = SingleRef<T>[]
-
-
-/**
- * A type that expresses a value is either an {@link LionWebId} or a value to indicate that resolution to a node previously failed.
- */
-export type IdOrUnresolved = LionWebId | typeof unresolved;
-
-
-/**
- * @return the serialization of the given {@link SingleRef single reference target}, as a {@link LionWebId LionWeb ID}.
- */
-export const serializedRef = <NT extends Node>(ref: SingleRef<NT>): LionWebId | typeof unresolved =>
-    ref === unresolved ? unresolved : ref.id
+export type MultiRef<NT extends Node> = SingleRef<NT>[]