+++ packages/delta-protocol-common/src/translators/event-to-delta.ts
@@ -116,7 +116,7 @@ export const eventToDeltaTranslator = (
     const eventAsDelta = (event: Event, idMapping: IdMapping): IDelta | undefined => {
 
         const deserializedNodeFrom = (chunk: LionWebJsonChunk): INodeBase => {
-            const { roots, idMapping: newIdMapping } = deserializeWithIdMapping(chunk, /* dependentNodes: */ [], idMapping)    // (deserializer should take care of installing delta receiver)
+            const { roots, idMapping: newIdMapping } = deserializeWithIdMapping(chunk, idMapping)    // (deserializer should take care of installing delta receiver)
             if (roots.length !== 1) {
                 throw new Error(`expected exactly 1 root node in deserialization of chunk in event, but got ${roots.length}`)
             }