+++ packages/class-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.8.0 — not yet released
+
+* Receive the `IdOrUnresolved` type from the `core` package, and export that and the `idFrom` function.
+* Improve the API surface of the `nodeBaseDeserializer[WithIdMapping]` functions:
+    * Expose a `DeserializerConfiguration` type, which encompasses the (partially optional) parameters that are likely unchanging per invocation of the deserializer.
+    * Use that type to as the single parameter to the `nodeBaseDeserializer[WithIdMapping]` functions.
+      The original signature is also still valid, but deprecated.
+    * Move the `propertyValueDeserializer` and `problem[s]Handler` parameters of the `Deserializer` type to the `DeserializerConfiguration` type.
+      *Note* that this is a **breaking change**, but these parameters weren’t used outside of tests anyway.
+* Simplify `Deserializer` type, by removing the `dependentNodes` parameter: (only) pass an appropriate instance of `IdMapping` instead.
+  *Note* that this is a **breaking change**, but either parameter is optional and carries the same information, so this should avoid confusion and misuse.
+
+
 ## 0.7.2
 
 * Implement a `propertyValueSerializerWith` function that produces a `PropertyValueSerializer` instance to serialize properties’ values with, properly dealing with enumerations as well.